chore(tileserver): clear clippy backlog + re-tighten CI gate
Fix the full -D-warnings clippy/rustc backlog across the workspace so the
strict gate can come back on:

- Real fixes: .clamp() over min/max, vec![] over Vec::new()+push,
  needless_range_loop → enumerate (aniso/unified/slope), FRAC_1_SQRT_2
  constant, zN → z_n, doc-list wording, empty-line-after-doc, Prefs struct
  init over default-then-reassign (route_plan), unused test bindings,
  removed the stale DemMeta bench.
- Scoped allows for genuinely-fine design lints: type_complexity +
  arc_with_non_send_sync + too_many_arguments crate-wide in pathfind/api
  (boxed cost closures, thread-local recorder Arcs, multi-knob solve entry
  points); dead_code / field_reassign in test scaffolding;
  assertions_on_constants on a const-bounds guard test.

CI: restore RUSTFLAGS=-D warnings and drop the clippy continue-on-error —
the gate is strict again. `cargo clippy --workspace --all-targets --
-D warnings` is clean.

+++ apps/tileserver/crates/turbo-tiles-admin/src/routes/tus.rs
@@ -453,7 +453,7 @@ fn free_disk_bytes(path: &Path) -> Option<u64> {
         if rc != 0 {
             return None;
         }
-        Some(u64::from(stat.f_bavail) * u64::from(stat.f_frsize))
+        Some(u64::from(stat.f_bavail) * stat.f_frsize)
     }
     #[cfg(not(unix))]
     {
